// UI/TADHACK/src/app/model/create-docket.model.ts
export class Case
    {
        description: string;
        media: string;
        caseType: string;
        capturerIdNo: string;
        caseStatus:string; 
        victimFullName: string;
        victimID: string;
        victimPassword: string;
        victimAddress: string;
        victimCellNo: string;
        victimGender: string;
      }